+++ tests/entrypoints/serve/rpc/test_collective_rpc.py
